fix: file support for electron renderer
